fixed multiple bugs related to std::map access
